[0.1.x] ~ int->long
git-svn-id: https://pykd.svn.codeplex.com/svn@73301 9b283d60-5439-405e-af05-b73fd8c4d996

+++ test/scripts/typedvar.py
@@ -174,7 +174,7 @@ class TypedVarTest( unittest.TestCase ):
 
         # if debug: g_unTypedPtrToFunction point to jmp EnumWindowsProc2 (e9 xxxxxxxx)
         self.assertTrue( ( target.module.offset("EnumWindowsProc2") == tv1 ) or
-                         ( 0xE9 == pykd.ptrByte( int(tv1) ) ) )
+                         ( 0xE9 == pykd.ptrByte( long(tv1) ) ) )
 
         tv2 = target.module.typedVar( "g_unTypedPtrToFunction" )
         self.assertEqual( tv1, tv2 )
